Add the other messages
