+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "Vizly-Notebook",
+	"name": "Thread",
 	"version": "0.1.0",
 	"private": true,
 	"scripts": {
@@ -74,16 +74,16 @@
 		"remark-breaks": "^4.0.0",
 		"remark-gfm": "^4.0.0",
 		"remark-math": "^6.0.0",
+		"shared-thread-utils": "file:./shared-thread-utils",
 		"socket.io": ">=2.4.0",
 		"socket.io-parser": ">=3.3.3",
-		"shared-thread-utils": "file:./shared-thread-utils",
 		"uuid": "^9.0.1",
 		"vscode-languageserver-protocol": "^3.17.5",
 		"xmlhttprequest-ssl": ">=1.6.2",
 		"xterm": "^5.3.0",
 		"zod": "^3.22.4",
 		"zundo": "^2.0.3",
-		"zustand": "^4.4.6"
+		"zustand": "^4.5.2"
 	},
 	"devDependencies": {
 		"@types/lodash.debounce": "^4.0.9",
