+++ shared-thread-utils/package.json
@@ -7,10 +7,17 @@
 		"build": "tsc"
 	},
 	"dependencies": {
-		"ai": "^3.0.19",
-		"openai": "^4.33.0"
+		"@sentry/nextjs": "^8.10.0",
+		"ai": "3.0.19",
+		"best-effort-json-parser": "^1.0.1",
+		"langfuse": "^3.11.2",
+		"lodash": "^4.17.21",
+		"openai": "^4.33.0",
+		"uuid": "^10.0.0"
 	},
 	"devDependencies": {
+		"@types/lodash": "^4.17.5",
+		"@types/uuid": "^9.0.8",
 		"@types/webpack": "^5.28.5",
 		"typescript": "^4.0.0"
 	},
